+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - rename `Downloader::get` to `url` **breaking**
 - rename `RequestBuilder::send` to `get` and change return value to `BytesResponse` **breaking**
 - rename `RequestBuilder::hash` to `verify_hash` **breaking**
+- add `RequestBuilder::calculate_hash`
 - add `RequestBuilder::save_to_file`
 
 ## 0.2.0 - 2025-11-21